requirements.txt
aiogram==3.7.*
asyncpg
python-dotenv
